# perf: クラス間連接コストの計算を高速化