--- a/Sources/KanaKanjiConverterModule/ConversionAlgorithms/Core/FullInputProcessing.swift
+++ b/Sources/KanaKanjiConverterModule/ConversionAlgorithms/Core/FullInputProcessing.swift
@@ -102,12 +102,13 @@ extension Kana2Kanji {
     }
     /// N-Best計算を高速に実行しつつ、遷移先ノードを更新する
     func updateNextNodes(with node: LatticeNode, nextNodes: some Sequence<LatticeNode>, nBest: Int) {
+        let ccLatter = self.dicdataStore.getCCLatter(node.data.rcid)
         for nextnode in nextNodes {
             if self.dicdataStore.shouldBeRemoved(data: nextnode.data) {
                 continue
             }
             // クラスの連続確率を計算する。
-            let ccValue: PValue = self.dicdataStore.getCCValue(node.data.rcid, nextnode.data.lcid)
+            let ccValue: PValue = ccLatter.get(nextnode.data.lcid)
             // nodeの持っている全てのprevnodeに対して
             for (index, value) in node.values.enumerated() {
                 let newValue: PValue = ccValue + value