* クラス間連接コストの計算は探索中頻繁に行われるが、ここが思ったよりも時間を食っていた。
* ccValueの取得においてメモリ上の排他制御が問題になっていたので、行単位アクセスに最適化したAPIを用いて改善した

これにより #249 で取り扱ったKanaKanjiConverterModuleWithDefaultDictionaryTests.ZenzaiTests/testGradualConversion_Roman2Kanaのケースでは10%の改善が得られた。一連の高速化の結果は以下の通り。

* #249 適用前: 10.46s
* #249 適用後: 7.69s
* #250 適用後: 6.71s
* #252 適用後: 5.99s